requirements.txt
xgboost
astropy
matplotlib
git+https://github.com/AstroHuntsman/gunagala.git
pandas
numpy
scipy
photutils
specutils
astroquery
pyyaml
pysynphot
xml
xmltodict
urllib
sklearn
datetime
netcdf4
ephem
healpy
pytest
